twitter/requirements.txt
twscrape>=0.12.0
python-dotenv>=1.0.0
pandas>=2.0.0
python-telegram-bot>=21.0